+++ client/client/tx.go
@@ -3,7 +3,7 @@ package client
 import (
 	"context"
 	"fmt"
-	relayerclient "github.com/babylonlabs-io/babylon/client/relayer"
+	"github.com/babylonlabs-io/babylon/client/babylonclient"
 	"sync"
 
 	signingv1beta1 "cosmossdk.io/api/cosmos/tx/signing/v1beta1"
@@ -26,10 +26,10 @@ import (
 )
 
 // ToProviderMsgs converts a list of sdk.Msg to a list of provider.RelayerMessage
-func ToProviderMsgs(msgs []sdk.Msg) []relayerclient.RelayerMessage {
-	relayerMsgs := make([]relayerclient.RelayerMessage, 0, len(msgs))
+func ToProviderMsgs(msgs []sdk.Msg) []babylonclient.RelayerMessage {
+	relayerMsgs := make([]babylonclient.RelayerMessage, 0, len(msgs))
 	for _, m := range msgs {
-		relayerMsgs = append(relayerMsgs, relayerclient.NewCosmosMessage(m, func(signer string) {}))
+		relayerMsgs = append(relayerMsgs, babylonclient.NewCosmosMessage(m, func(signer string) {}))
 	}
 	return relayerMsgs
 }
@@ -47,7 +47,7 @@ func (c *Client) SendMsgsToMempool(ctx context.Context, msgs []sdk.Msg) error {
 	if err := retry.Do(func() error {
 		var sendMsgErr error
 		krErr := c.accessKeyWithLock(func() {
-			sendMsgErr = c.provider.SendMessagesToMempool(ctx, relayerMsgs, "", ctx, []func(*relayerclient.RelayerTxResponse, error){})
+			sendMsgErr = c.provider.SendMessagesToMempool(ctx, relayerMsgs, "", ctx, []func(*babylonclient.RelayerTxResponse, error){})
 		})
 		if krErr != nil {
 			c.logger.Error("unrecoverable err when submitting the tx, skip retrying", zap.Error(krErr))
@@ -66,21 +66,21 @@ func (c *Client) SendMsgsToMempool(ctx context.Context, msgs []sdk.Msg) error {
 // ReliablySendMsg reliable sends a message to the chain.
 // It utilizes a file lock as well as a keyring lock to ensure atomic access.
 // TODO: needs tests
-func (c *Client) ReliablySendMsg(ctx context.Context, msg sdk.Msg, expectedErrors []*errors.Error, unrecoverableErrors []*errors.Error) (*relayerclient.RelayerTxResponse, error) {
+func (c *Client) ReliablySendMsg(ctx context.Context, msg sdk.Msg, expectedErrors []*errors.Error, unrecoverableErrors []*errors.Error) (*babylonclient.RelayerTxResponse, error) {
 	return c.ReliablySendMsgs(ctx, []sdk.Msg{msg}, expectedErrors, unrecoverableErrors)
 }
 
 // ReliablySendMsgs reliably sends a list of messages to the chain.
 // It utilizes a file lock as well as a keyring lock to ensure atomic access.
 // TODO: needs tests
-func (c *Client) ReliablySendMsgs(ctx context.Context, msgs []sdk.Msg, expectedErrors []*errors.Error, unrecoverableErrors []*errors.Error) (*relayerclient.RelayerTxResponse, error) {
+func (c *Client) ReliablySendMsgs(ctx context.Context, msgs []sdk.Msg, expectedErrors []*errors.Error, unrecoverableErrors []*errors.Error) (*babylonclient.RelayerTxResponse, error) {
 	var (
-		rlyResp     *relayerclient.RelayerTxResponse
+		rlyResp     *babylonclient.RelayerTxResponse
 		callbackErr error
 		wg          sync.WaitGroup
 	)
 
-	callback := func(rtr *relayerclient.RelayerTxResponse, err error) {
+	callback := func(rtr *babylonclient.RelayerTxResponse, err error) {
 		rlyResp = rtr
 		callbackErr = err
 		wg.Done()
@@ -95,7 +95,7 @@ func (c *Client) ReliablySendMsgs(ctx context.Context, msgs []sdk.Msg, expectedE
 	if err := retry.Do(func() error {
 		var sendMsgErr error
 		krErr := c.accessKeyWithLock(func() {
-			sendMsgErr = c.provider.SendMessagesToMempool(ctx, relayerMsgs, "", ctx, []func(*relayerclient.RelayerTxResponse, error){callback})
+			sendMsgErr = c.provider.SendMessagesToMempool(ctx, relayerMsgs, "", ctx, []func(*babylonclient.RelayerTxResponse, error){callback})
 		})
 		if krErr != nil {
 			c.logger.Error("unrecoverable err when submitting the tx, skip retrying", zap.Error(krErr))
@@ -146,9 +146,9 @@ func (c *Client) ReliablySendMsgs(ctx context.Context, msgs []sdk.Msg, expectedE
 
 // ReliablySendMsgsWithSigner reliably sends a list of messages to the chain.
 // It utilizes the signer private key to sign all msgs
-func (c *Client) ReliablySendMsgsWithSigner(ctx context.Context, signerAddr sdk.AccAddress, signerPvKey *secp256k1.PrivKey, msgs []sdk.Msg, expectedErrors []*errors.Error, unrecoverableErrors []*errors.Error) (*relayerclient.RelayerTxResponse, error) {
+func (c *Client) ReliablySendMsgsWithSigner(ctx context.Context, signerAddr sdk.AccAddress, signerPvKey *secp256k1.PrivKey, msgs []sdk.Msg, expectedErrors []*errors.Error, unrecoverableErrors []*errors.Error) (*babylonclient.RelayerTxResponse, error) {
 	var (
-		rlyResp     *relayerclient.RelayerTxResponse
+		rlyResp     *babylonclient.RelayerTxResponse
 		callbackErr error
 		wg          sync.WaitGroup
 	)
@@ -208,9 +208,9 @@ func (c *Client) SendMessageWithSigner(
 	ctx context.Context,
 	signerAddr sdk.AccAddress,
 	signerPvKey *secp256k1.PrivKey,
-	relayerMsgs []relayerclient.RelayerMessage,
+	relayerMsgs []babylonclient.RelayerMessage,
 ) (result *coretypes.ResultBroadcastTx, err error) {
-	cMsgs := relayerclient.CosmosMsgs(relayerMsgs...)
+	cMsgs := babylonclient.CosmosMsgs(relayerMsgs...)
 	var (
 		num, seq uint64
 	)
@@ -481,11 +481,11 @@ func Sign(
 // - we do not support cancellation of submitting messages
 // - the only timeout is the block inclusion timeout i.e block-timeout
 // TODO: To properly support cancellation we need to expose ctx in our client calls
-func (c *Client) InsertBTCSpvProof(ctx context.Context, msg *btcctypes.MsgInsertBTCSpvProof) (*relayerclient.RelayerTxResponse, error) {
+func (c *Client) InsertBTCSpvProof(ctx context.Context, msg *btcctypes.MsgInsertBTCSpvProof) (*babylonclient.RelayerTxResponse, error) {
 	return c.ReliablySendMsg(ctx, msg, []*errors.Error{}, []*errors.Error{})
 }
 
-func (c *Client) InsertHeaders(ctx context.Context, msg *btclctypes.MsgInsertHeaders) (*relayerclient.RelayerTxResponse, error) {
+func (c *Client) InsertHeaders(ctx context.Context, msg *btclctypes.MsgInsertHeaders) (*babylonclient.RelayerTxResponse, error) {
 	return c.ReliablySendMsg(ctx, msg, []*errors.Error{}, []*errors.Error{})
 }
